Moving tests to LanceLakeWriterTest
